LP# 1337550 remove schema.org itemOffered, add price property
We use the "offers" property to link the Product to the Offer, so there
is no reason to use a reciprocal "itemOffered" property to link back to
the Product from the Offer. The circular reference confuses some tools
(http://rdfa.info/play) and may complicate general processing, so let's
simply remove it.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>