Fix currently harmless but still confusing sigil error
get_org_descendants() returns an array ref, not a list, so it doesn't
make sense to assign it to an '@orgs' variable.  Lucky for us, our
where-parser treats this single element list containing the ref in the
same manner as the proper ref itself, but at best, it is still very
confusing to see.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>