LP1282277_LP1282286_Unitialized_Vars_FIx
The commit to prompt users if they are going to lose data in MARC edit
added some new variables (tabs, and tab) in various files.  If the MARC
edit window does not appear in a tab, then these variables remain
unitialzed and cause problems for MARC edits that appear in their own
window.  By checking for the existence of the variables before MARC
edito loads in its own windown, the problems are resolved.

Signed-off-by: Liam Whalen <liam.whalen@bc.libraries.coop>
Signed-off-by: Jennifer Pringle <jpringle@sitka.bclibraries.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>