LP#1334693 make_release optional -j <osrf_js_path>
make_release supports a "-j" option for passing in the OpenSRF JS path:

make_release [options] -j /openils/lib/javascript

This is useful when running make_release, specifically the staff client
building portion, on a machine where the OpenSRF libs are available
(say, from a git checkout) but osrf_config is not available because
OpenSRF is not actually installed.

The script will attempt to find osrf_config and the JS path
automatically, so under normal circumstances, -j is not required.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>