LP#1346421 TPAC copy view / edit links
Adds "view" and "edit" links next to each copy in the TPAC copy grid
when in embedded mode.  The "edit" link only appears when the user has
permission to edit the specified copy.

Via the XUL volume_item_creator, the ui.unified_volume_copy_editor org
unit setting is inspected to determine which UI to load.  Also, the data
needed to load the subsequent UI is fetched within the
volume_item_creator, so the catalog doesn't have to.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>