LP#1350371 PO name on create w/ dupe detect
Adds support for entering a PO name at PO creation time, both from the
picklist page and from the PO create page.  If a name is entered which
matches the name of an existing PO at or below the ordering agency, a
link to the matching PO is shown and the user is prevented from creating
the PO until the name is changed or cleared.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>