LP#1350371 display existing copies
In the picklist / PO lineitem list view, display the number of existing
catalog copies owned at or below the PO ordering agency or Picklist org
unit whose status is not in some form of lost or missing.

Value is displayed beside the lineitem ID and is styled bold/red if the
count is greater than zero.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>