LP#1304462: Add schema.org copyrightYear, manufacturer for RDA 264 fields
It looks like the best we can do is to define a "manufacturer" property (and
corresponding Organization entity) when 264 ind2="3"; and we can provide a
copyrightYear when we have 264 ind2="4".

Even then, MARC / RDA cataloging practices end up giving us copyright notices
like '@2012' instead of actual dates, so we have to regex the year out of the
notice to satisfy the Integer range for copyrightYear. Theoretically we'll be
in trouble if the copyright notice includes some other 4-digit number (for
example, a multipart work catalogued in a single record that has a date range),
but what are you gonna do? It's MARC / RDA.

Same for manufacturer name, which seems likely (if the LoC examples that Yamil
provided are any guide) to include statements like "Manufactured by Foobar"
instead of cleanly structured data. The RDA fields were added to MARC21 in
2011, you would have thought that the benefits of providing clean data would
have been obvious by then. Grumble grumble grumble. Yes, this is meant to be
a transcription field, and theoretically one could use a 7xx to record the
publisher/distributor/manufacturer/producer in a more cleanly stated factual
way that would lend itself to actual linked data (hey, maybe even providing a
link to an external resource that could act as an identifier), but NOBODY DOES
THAT.

So we do the best we can and soldier on.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>