LP#1350042 detect is-staff mode for browser client
The browser client does not pass an OILS-Wrapper header to the embedded
catalog (etc.) so consider the presence of a workstation an indication
that the TPAC should operate in staff mode.

NOTE: this may require some additional thought, since this means all
TPACs within an authenticated browser will run in staff mode, which may
not be desired.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>