LP#1329503 avoid run now + later combined schedule
When editing a report, ensure only one of "now" vs. "schedule for later"
options is selected by default.  Similarly, if the schedule option is
activated by default, ensure the date fields are enabled.

If the run time of the most recent report is in the past, default to the
current date for the schedule instead of the no longer valid date.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>