LP#1359934: prevent error opening patron search form
As a result of the patch for OpenSRF bug 1316245, JSON2js
in the OpenSRF library is now stricter -- in particular, passing
an undefined value to it results in an exception being thrown.

This patch runs with the increased strictness and fixes an
issue where the patron search form was passing an undefined value
to JSON2js.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>