LP#1362210: Install PostgreSQL packages where we can
Rather than relying on purely manual installation, carve out a common
set of packages that are needed by both the Evergreen server and the
standalone PostgreSQL server and add them as dependencies to both
scenarios. In the worst case, the package manager will see a package
that it has already installed and skip it (if the database server is on
the same OS instance as the Evergreen server).

It sure beats upgrading your database server and wondering why things
like the addition of call numbers fails with database query errors when
it turns out that you're missing the Library::CallNumber::LC Perl module
:)

It also makes the README shorter and requires fewer steps for users to
follow to install successfully. Win win win?

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>