LP#1358392: See references not always displaying on browse search
When browsing for an unauthorized heading, See references are not always
displaying on browse search. Those that do display surface due to the
existence of cross-reference(s).

Desired behaviour is that unauthorized headings should also surface in browse
results, even without the existence of cross-reference(s), as long as the
authority record containing the unauthorized heading controls at least one bib.
There are many authority records that do not have cross-references but do contain
unauthorized headings, so browse results that would otherwise be hidden would now
be exposed.

This fix adds an additional logic to check if the browse term is an
unauthorized term and if so, perform a slightly modified procedure
but otherwise, perform the usual procedure so as not to break any existing
functionality.

Signed-off-by: Srey Seng <sreyseng@gmail.com>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>

Conflicts:
	Open-ILS/src/sql/Pg/030.schema.metabib.sql

Rebased to account for eliminated staged search work.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>