LP#1198465 Allow fine generator to respect a stop_fines filter
In order to support situations where billings should be generated on
specific circs, even when stop_fines is set, we allow a filter to
be passed.  The default of not passing a filter is the same as requiring
that stop_fines be NULL, which is the pre-commit behavior.

We also remove a never-used parameter.

Signed-off-by: Mike Rylander <miker@esilibrary.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>