LP#1306814: Make use of patron timeout setting for selfcheck
As described in the bug, the library setting "Self Check: Patron Login Timeout
(in seconds)" does not appear to work with the newer selfcheck interface.

It looks like some of it was already being pulled in, like the variable, but
the interface was not utilizing it.

This commit borrows from the old JS selfcheck and implements a default of
three minutes for selfcheck timeout otherwise handled by the library setting.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Michele Morgan <mmorgan@noblenet.org>