LP#1444644 Copy Import Development Work
Two new options for importing holdings have been added to MARC Batch
Import/Export:

1) Auto-overlay On-order Cataloguing Copies

This is similar to "Auto-overlay In-process Acquisitions Copies," but
for copies that were not created from an acquisitions workflow.
Holdings information in the incoming record will be used to overlay any
existing On Order copies for the matching record which belong to the
owning library defined in the Holdings Import Profile.  The Holdings
Import Profile is also used to match incoming to existing copies, if
possible; otherwise, On Order copies are overlaid in the order they were
created.  The call number will also be overlaid if the incoming record
provides one.

2) Use Org Unit Matching in Copy to Determine Best Match

When there are multiple potential matching records, this feature allows
the user to automatically select the record which has the most copies at
libraries near the importing library in the org tree.  That is, starting
at the importing library, it climbs the org tree, gradually expanding
the scope at which it checks for holdings on matching records; once
holdings are found, the record with the most holdings at that scope is
selected for overlay.  If there are no matching records with holdings,
then the default best match overlay is attempted.

Two new permissions are added to control the use of these new features.

Signed-off-by: Liam Whalen <liam.whalen@bc.libraries.coop>
Signed-off-by: Jeff Davis <jeff.davis@bc.libraries.coop>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>