LP#1378169: make context OU selector on funds page sticky
This patch uses local storage to persist the selected value of the
context OU selector on the funds page.  This is meant to ease
workflows where an acquisitions administrator needs to frequently
transfer money between funds that are attached to children of
a system-level OU.  Specifically, rather than defaulting to
the user's workstation OU, the selector will default to the
OU selected the last time the funds page was opened.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>