LP#1379815 Add missing behavior to update_copy_stat_entries
The current function doesn't provide the level of control we need, even
with the 'delete_stats' flag set to zero. This commit adds a new
'add_or_update_only' option which preserves existing stat cat entry maps
for any stat cat not represented on the incoming copy object, and
updates values when appropriate.

This commit also updates the function comment for better clarity.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Ben Shum <bshum@biblio.org>