LP#957466: Update editor/edit_date/source on overlay
This commit does three things.

1. It removes some declared variables that are never used in this
function.

2. It updates the bib record edit_date field (along with the editor) if
an editor is found in the MARC. If an editor is not found (or doesn't
match an Evergreen user), it seems best to leave the edit_date
unchanged so as not to imply that the previous editor is responsible for
the newest edit.

3. If a bib source is chosen in the vandelay importer UI, it updates the
bib record with the source. To access this field, the reference to table
"queued_record" is replaced by its child table "queued_bib_record".
Since the new table is a child of the other, all of the other needed
values are still available.

Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Martha Driscoll <driscoll@noblenet.org>
Signed-off-by: Dan Wells <dbw2@calvin.edu>