LP#1390138: Updated upgrade docs for 2.7.1.
Added in section about checking the release notes for other upgrade steps to the end.

Added in the step to restart memcached as Galen Charlton mentioned in his 11/5 dev list email
about clearing out old login sessions for security reasons.

Added in Jessie and Trusty with a pointer to the Ruby EDI bug that users of the distributions
will run into.

Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Ben Shum <bshum@biblio.org>