LP#1384740 marc_stream_importer auth and cleanup
Add support for importing authority records to marc_stream_importer and
clean up some cruft along the way.  A single instance of the script can
import either type of record.  The record leader is inspected to
determine if it's an authority record or not.

New options:

--auth-merge-profile
--auth-queue
--auth-source
--bib-import-no-match
--bib-auto-overlay-exact
--bib-auto-overlay-1match
--bib-auto-overlay-best-match
--auth-import-no-match
--auth-auto-overlay-exact
--auth-auto-overlay-1match
--auth-auto-overlay-best-match

Deprecated options:

--import-no-match
--auto-overlay-exact
--auto-overlay-1match
--auto-overlay-best-match

These map to the new --bib-* versions for now, but will be removed
in a future version for clarity / consistency.  The script will
print a deprecation warning when these are used.

Cleanup:

--import-by-queue is no longer supported.  This option serves no
particular purpose and is a bad idea when re-using the same queue over
and over as most people do, because queue bloat will increase run times.

--noqueue (AKA "direct import") is no longer supported.  All imports go
through Vandelay now.

Opens a new XMPP connection with each forked child during child init.
Previously, a single XMPP connection was shared by all forked children,
which lead to crossed streams and chaos.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>