LP#1396161: Improve public docs and change osrf-config default from script help
The public docs have an "Action Triggers" chapter, so this commit
applies the same improvements there to match the help docs in the
script. Also in the public docs, in two places a filename was split
across lines, causing a gap to appear in the HTML version of the docs.
That is corrected here.  This commit also removes the 'osrf-config'
option from the examples in the script help docs, since they are using
the default value.

Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Ben Shum <bshum@biblio.org>