LP#1198465 Allow null fields to stay null in to_fieldmapper()
In CDBI.pm, we can convert DB objects to Fieldmapper objects using
to_fieldmapper().  Part of the conversion was forcing stringification,
which does what we need for nested objects, but also converts 'nulls'
(undefineds) to 'empty' values.  This causes problems when trying to
write these Fieldmapper objects back to the DB using cstore (or the
like).

This commit simply lets undefined values stay that way, which in turn
prevents NULLs from trying to become empty strings in the DB.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Bill Erickson <berickxx@gmail.com>