LP#1198465 Recombine generate_fines()
The generate_fines() method was split back in 2010.  There is some
concern about the speed of checkin (e.g. via SIP for a sorter, where we
don't care about fines), and this may have been the reason for the
split.  It might also have been the fact that waiting for the storage
xact to finish was causing the circ xact to timeout.

In any case, since we are bringing them into the same xact now, reason
number 2 is invalid, and reason number 1 won't be helped by the split
anymore (since there is no longer any parallel action).  Let's bring it
back together.

Ultimately, as suggested by Mike, we will probably want to refactor a
bit and return a streaming response for sorting location, then a final
response once we have the fine generation figured out.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Bill Erickson <berickxx@gmail.com>