LP#1400100 - Avoid extra </div> when OpenURL is enabled
When the OpenURL resolver is enabled, the record summary.tt2 page always
inserts a </div>, whether or not there actually are any OpenURL results to
point at (which is the condition that triggers the opening <div>).

This has the effect of resulting in the generated RDFa not including anything
below the OpenURL section, effectvely preventing the schema.org/Offer types
from being associated with the record. Which is not good for the semantic web
of library data.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>