LP#980296: Update void on claims returned for longoverdue status.
We add settings for voiding longoverdue fees on claims returned.

We modify code in Circ.pm to check those settings and void those
bills when a circulation is claims returned.

We also make some logical repairs to the new void_lost function
in CircCommon.pm and to event handling where that function is
called in set_circ_claims_returned in Circ.pm.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>