LP#1373594: Also ignore subfield 7 in get_graphics_880s
In the Czech Republic the subfield 7 of MARC fields 1XX and 7XX are used for
the National authority control number.

The problem is, that in current Evergreen versions, the subfield 7 is displayed
as a part of the link to author seach in search result or record details, but
when using the linkt to search rcords of the same author, no results are
returned, no matter whether the records exist or not.

That is why we nedd to add the subfield 7 into get_graphic_880s.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>