Add comment to run 2.6.2-2.6.3 upgrade script if missed during 2.6-2.7 major upgrade
As discovered by Dan Scott, the 2.7.0 and 2.7.1 shipped with a missing upgrade
script for 2.6.2-2.6.3 and if one upgraded directly from 2.6.2 to 2.7.0, you
would miss out on the fixes contained within.

Add a comment to the end of the 2.7.2 upgrade script to tell folks to go back
and run the 2.6.2-2.6.3 upgrade (which is now included in the tarball and
branches for 2.7 series).

Signed-off-by: Ben Shum <bshum@biblio.org>