Docs: fix asciidoc errors
When https://bugs.launchpad.net/evergreen/+bug/1386854 was merged to add a new
docs section for "Locally hosted content", it looks like there was a bad git
conflict resolution and an extra block of ======= was left behind.

Remove this to fix headings and table of contents for the documentation site.

Signed-off-by: Ben Shum <bshum@biblio.org>