LP#1406025: Provide graceful human output for HTTP errors
For the expected HTTP errors of HTTP_GONE (record deleted)
and HTTP_NOT_FOUND (record never existed), return the HTML
pages that we used to return (red border warning that the
record was deleted, or just an empty template for
non-existing records) so that the user can at least try
another search. However, the HTTP status code gets set to
410 or 404 as expected so that machines can react
accordingly.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>