LP#1407507: Update relator codes
Per bug # 1046149, we added a script that will generate the contents of the
opac/parts/relators.tt2 file to provide more granular credits for added entries
that have relator codes. However, we haven't run that script since 2013, and
as there are a number of new relator codes, along with some revisions to the
existing relator terms, we should update relators.tt2 accordingly.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>