LP#1409844: Towards more meaningful catalogue <title> elements
Currently the catalogue <title> elements are formed by concatenating the page
title to "Catalog - ". This means that on browsers with more than a few tabs
open, users will only see "Catalog -", which doesn't help them navigate their
tabs. Also, in search results in general search engines, "Catalog - " is
expected to be a significant word as it appears first in the title of the page.

Instead, let's remove the "Catalog - " prefix entirely and add "- <library
name>" as a suffix to the page title. That way, a user will see the page title
in bookmarks, search results, and other contexts with the most important
information up front (whether that's the title of the work in record details,
or the fact that it's a search for some string), followed by the name of the
library to give it a specific context.

In addition, use the name of the library in the OpenSearch definition to be
more meaningful than the current hardcoded "Evergreen OpenSearch" (many users
will not even know what Evergreen is, if they stumble across the little "Add
search engine" widget at all).

And finally, use the HTML and URI filters appropriately to avoid any possible
damage :)

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>