LP#1365661 Found patrons with same name not opening
This patch fixes the error, while registering a new patron
and clicking the link "Found {$0} patron(s) with the same
name" did not previously work in the staff client.

To reproduce the bug, register a new patron with the name of an
existing patron.  The little widget with the message described
above will appear.  The message is a hyperlink, but nothing
appears to happen if you click it.  The JavaScript console will
contain an error ReferenceError:js2JSON is not defined.

After applying this patch, clicking the hyperlink does a search
for matching patrons as intended.  The ReferenceError also
disappears.

Signed-off-by: Adam Bowling <abowling@emeralddata.net>
Signed-off-by: Terran McCanna <tmccanna@georgialibraries.org>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>