LP #1406350 Mobile Device Navigation Issue Fix for Shelf Browser
Currently, when a browser window is narrowed too much, whether
in a mobile browser, or on a desktop or laptop machine, the
"Previous Page" and "Next Page" navigation links disappear. The
only solution that seems to solve this is by removing the
"THEAD" element from the shelf browser table.

This patch removes that tag, adds a new row to imitate a new
border as existed before, and modifies the CSS in .../opac/
css/style.css.tt2 to address that difference.

Signed-off-by: Adam Bowling <abowling@emeralddata.net>
Signed-off-by: Yamil Suarez <yamil@yamil.com>
Signed-off-by: Ben Shum <bshum@biblio.org>