LP#1412893: Only apply lost and paid status with the proper transactions
As discovered while testing the lost and paid feature, when closing earlier transactions where
the item is currently lost, it could change the item status to lost and paid prematurely.

Change the logic so that it only does this if the current transaction being closed either has
a status of lost or longoverdue.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>