LP#1244547 Maintain Filter Settings on Library Settings Editor
After applying a filter to the rows on the Library Settings
Editor, and then selecting a different OU, the filter settings
are not considered unless the user clicks the "Filter" button
once again.

This patch maintains the filter settings on the edior until
the user clicks the "Clear Filter" button.

Signed-off-by: Adam Bowling <abowling@emeralddata.net>
Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>