LP#1413592: Don't change lost/long-overdue fines/fees on zero-balance
The situation that has been reported as the worst offender for creating
negative balances for folks that do not want that to happen goes like this:

1) Item is lost or goes long-overdue
2) Fees/fines charged
3) Said fees/fines are paid in full
4) Item is found and returned
4a) "Void fines and fees on lost item return" settings are enabled
4b) "Reinstate overdue fines on lost item return" setting is enabled
4c) System dutifully voids the lost fines/fees
4d) System dutifully reinstates the overdue fines

So we add an YAOUS to ensure that when a lost item is returned and the
balance of the transaction is exactly $0, the lost fines and fees are not
voided EVEN IF "void fines and fees on lost/long-overdue return" settings are
enabled, and voided overdue fees are likewise not reinstated EVEN IF
"restore overdue on lost/long-overdue return" settings are enabled.

The reason for the second condition is that the point of this new setting is
to have Evergreen consider zero-balance lost/long-overdue transactions as
"handled", and that they should not automatically change.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>