LP#1415898 Fix Potential Cstore Starvation in KPAC
The getit_results function in EGKPacLoader.pm does xact_begin, which
opens a connected session to a cstore drone.  It later does xact_rollback,
which does not disconnect the session.  This leaves a drone hanging for
several seconds unable to responde to new requests.  This commit switches
to a rollback call that does disconnect the session.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Bill Erickson <berickxx@gmail.com>