LP#1415898 Fix Potential Cstore Starvation in the OPAC
When a user views their current circulations and/or renews them via
the OPAC, a connection is made to a cstore drone so that the user's
circulations can be retrieved from the database, bypassing the cache.
This connection is not closed, and thus a cstore drone is left unable
to respond to new requests for several seconds.  On a busy server,
this has actually caused the number of cstore drones to hit the limit
configured in opensrf.xml.  If that situation continues for too long,
it leads to the load on the server spiking to the point where services
must be restarted.

This commit changes from xact_rollback to rollback so that the connection
will be closed.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Bill Erickson <berickxx@gmail.com>