LP#1413592: Avoid generating fines in "consider zero balance closed" mode
We will now also disable the behavior triggered by the "Lost Checkin
Generates New Overdues" org setting when the new "Do not change fines/fees
on zero-balance LOST transaction" org setting is enabled.  This is to follow
the new rule that the new setting creates: a zero-balance LOST transaction is
considered handled, and should not be changed by the system.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>