LP#1425191 Restore summarization of serial units
In a previous commit, code was added for force summaries to be limited
to a certain type (to prevent cross type contamination of summaries).
This had the negative side-effect of preventing summaries on units, as
the units were not requesting a 'type' for their summary.

While we could force the unit code to send a type, it seemed generally
more flexible to restore the removed support for returning all summary
information in the absense of a requested type.  This could lead to
cross-type summaries for units, but that is a larger issue altogether
(i.e. how we can properly support cross-type units at all).

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>