LP#957466 Vandelay set the 905$u to current user if not defined
This version only adds a 905$u if none already exist in the document. It
does so by creating a new 905 tag with subfield u. It does not add on
to any possibly existing 905 tag that does not have a subfield u.

This will cause the code in the previous commit to trigger and update the
appropriate fields in biblio.record_entry.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Dan Wells <dbw2@calvin.edu>