LP#1282286 Tab is NULL error
When editing a MARC record in Vandelay or MFHD holdings in a popup XUL
WIndow, a tab is NULL error is presented to the user.

This happens because an undefined variable is being used when it should
not be.  This code adds a check for that variable to ensure it exists
before using it.  The variable used only matters in cases where tabs are
present, which is not the case with pop up MARC edit windows, so
ignoring the variable in these cases is fine.

Signed-off-by: Liam Whalen <liam.whalen@bc.libraries.coop>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Dan Wells <dbw2@calvin.edu>