LP#1379815 Add pgTAP test
This commit adds a pgTAP test which ensures that imported items are
being successfully added to the 'import_item' table. The test confirms
that the correct number of items have stat cat data in the table. The
test also ensures that the new xpath function works properly, simply by
using it.

Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Ben Shum <bshum@biblio.org>